requirements.txt
requests>=2.31.0
pyyaml>=6.0
jsonschema>=4.21.0
pandas